damage and heatl to float

+++ Assets/Scripts/Gameplay/Combat/Contexts/AfterHitContext.cs
@@ -14,7 +14,7 @@ namespace Project.Gameplay.Gameplay.Combat.Contexts
         public GridPosition TargetPosition { get; set; }
         public BoardGrid Grid { get; set; }
         
-        public int DamageDealt { get; set; }
+        public float DamageDealt { get; set; }
         public bool TargetDied { get; set; }
         public bool WasCritical { get; set; }
